requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.7
packaging==23.1
Pillow==9.5.0
pydantic==1.10.7
pytesseract==0.3.10
requests==2.30.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
